+++ pkg/codegen/testing/test/testdata/plain-schema-gh6957/go/xyz/staticPage.go
@@ -47,7 +47,7 @@ type staticPageArgs struct {
 
 // The set of arguments for constructing a StaticPage resource.
 type StaticPageArgs struct {
-	Foo *Foo
+	Foo *FooArgs
 	// The HTML content for index.html.
 	IndexContent pulumi.StringInput
 }
@@ -137,6 +137,16 @@ func (o StaticPageOutput) ToStaticPageOutput() StaticPageOutput {
 
 func (o StaticPageOutput) ToStaticPageOutputWithContext(ctx context.Context) StaticPageOutput {
 	return o
+}
+
+// The bucket resource.
+func (o StaticPageOutput) Bucket() s3.BucketOutput {
+	return o.ApplyT(func(v *StaticPage) s3.BucketOutput { return v.Bucket }).(s3.BucketOutput)
+}
+
+// The website URL.
+func (o StaticPageOutput) WebsiteUrl() pulumi.StringOutput {
+	return o.ApplyT(func(v *StaticPage) pulumi.StringOutput { return v.WebsiteUrl }).(pulumi.StringOutput)
 }
 
 type StaticPageArrayOutput struct{ *pulumi.OutputState }